Mechanical compiler-warning cleanup
No behavior change intended anywhere in this commit -- purely
eliminating warning categories:

- -Wignored-qualifiers: dropped the meaningless leading const on
  by-value FeatureType/MatcherType returns (Feature::getType/
  getMatcherType and all 8 subclass overrides). By far the largest
  single warning count in the original audit.
- -Wreorder: reordered constructor initializer lists to match member-
  declaration order in FeatureMatcher, Frame, KeyFrame, LocalMapping,
  LoopClosing, MapPoint, Sim3Solver, System, Tracking, Vocabulary --
  spot-checked each for an initializer depending on a
  later-in-the-list member; found none.
- -Wunused-parameter: dropped parameter names (kept types) on stub/
  not-yet-implemented bodies across FeatureExtractor.h,
  FeatureMatcher_superglue.cpp, System.cc's TrackStereo/TrackRGBD
  stubs, KeyFrame::UnprojectStereo; [[maybe_unused]] on Optimizer::
  LocalBundleAdjustment's pbStopFlag (its early-stop checks are
  already commented out -- restoring that logic is a separate,
  deliberate decision, not part of a warnings pass).
- -Wsign-compare: int loop counters compared against .size() changed
  to size_t, in Initializer.cc, FeatureMatcher.cpp, FeatureExtractor.
  cpp, Viewer.cc; FeatureMatcher.h's min_match_keyframes to size_t.
- -Wdeprecated-copy: added Frame::operator= (delegates to the existing
  copy constructor via destroy + placement-new) -- Frame is copy-
  assigned every frame in Tracking::GrabImageMonocular, and only had
  a custom copy constructor with no matching assignment operator.
- -Wunused-function/-Wunused-variable: moved Feature_orb32.h's ORB
  descriptor/orientation helpers (computeOrbDescriptor, bit_pattern_
  31_, etc.) into Feature_orb32.cpp -- they're private to that
  translation unit, but living in the header meant every other .cpp
  that included it got its own unused copy. Also dropped assorted
  dead locals (FeatureExtractor.cpp's unused timing code, LocalMapping
  ::CreateNewMapPoints' cosParallaxStereo/sinThr, ::KeyFrameCulling's
  featType, Viewer.cc's unused trajectoryCenter0/cameraHeight0 and an
  unused CreatePanel() return value).

+++ include/FeatureMatcher.h
@@ -85,7 +85,7 @@ public:
     // Matches two keyframes across all feature types to find candidate pairs for triangulation.
     // Reuses cached matches if available; otherwise runs parallel descriptor matching + LMEDS filtering.
     // Used in: LocalMapping::CreateNewMapPoints
-    const int min_match_keyframes = 10;
+    const size_t min_match_keyframes = 10;
     map<FeatureType, vector<pair<size_t,size_t>>> match_keyframes(const Keyframe& keyframe1, const Keyframe& keyframe2,
                                 const vector<FeatureType>& feat_types);
     void match_keyframes_for_triangulation(const Keyframe& keyframe1, const Keyframe& keyframe2,
